Drawing Circles to the screen seems finished.

+++ FirstSDLWindow/Game.h
@@ -36,6 +36,6 @@ private:
 	int flags;
 	bool running;
 
-	void DrawCircle(SDL_Renderer* renderer, int32_t centreX, int32_t centreY, int32_t radius);
+	void DrawCircle(SDL_Renderer* renderer, float centreX, float centreY, float radius);	// Draws a circle. Set the desired color before you call this function (circle rasterization)
 };
